// frontend/src/types/companies/CompanyContactInformationForm.type.ts
export interface CompanyContactInformationForm {
    email_company: string | null;
    address: string | null;
    phone_number: string | null;
    contact_email: string | null;
    hr_mail: string | null;
    website: string | null;
    facebook: string | null;
    instagram: string | null;
    twitter: string | null;
    linkedin: string | null;
    youtube: string | null;
    name: string;
    id: number;
}